fix(cpp_utils/windows_app_tools)!: 修复编译警告

+++ include/cpp_utils/windows_app_tools.hpp
@@ -41,7 +41,7 @@ namespace cpp_utils
             if ( size_needed <= 0 ) {
                 return std::pmr::wstring{ resource };
             }
-            std::pmr::wstring result{ size_needed - 1, '\0', resource };
+            std::pmr::wstring result{ static_cast< std::size_t >( size_needed - 1 ), '\0', resource };
             MultiByteToWideChar( Charset, 0, str, -1, result.data(), size_needed );
             return result;
         }
